+++ packages/util-crypto/src/nacl/sign.ts
@@ -5,7 +5,7 @@
 import { Keypair } from '../types';
 
 import nacl from 'tweetnacl';
-import { assert } from '@polkadot/util';
+import { assert, u8aToU8a } from '@polkadot/util';
 import { isReady, ed25519Sign } from '@polkadot/wasm-crypto';
 
 /**
@@ -22,10 +22,12 @@ import { isReady, ed25519Sign } from '@polkadot/wasm-crypto';
  * naclSign([...], [...]); // => [...]
  * ```
  */
-export default function naclSign (message: Uint8Array, { publicKey, secretKey }: Partial<Keypair>): Uint8Array {
+export default function naclSign (message: Uint8Array | string, { publicKey, secretKey }: Partial<Keypair>): Uint8Array {
   assert(secretKey, 'Expected valid secretKey');
 
+  const messageU8a = u8aToU8a(message);
+
   return isReady()
-    ? ed25519Sign(publicKey as Uint8Array, (secretKey as Uint8Array).subarray(0, 32), message)
-    : nacl.sign.detached(message, secretKey as Uint8Array);
+    ? ed25519Sign(publicKey as Uint8Array, (secretKey as Uint8Array).subarray(0, 32), messageU8a)
+    : nacl.sign.detached(messageU8a, secretKey as Uint8Array);
 }