+++ packages/util-crypto/src/schnorrkel/verify.ts
@@ -4,12 +4,17 @@
 
 import '../polyfill';
 
+import { u8aToU8a } from '@polkadot/util';
 import { sr25519Verify } from '@polkadot/wasm-crypto';
 
 /**
  * @name schnorrkelVerify
  * @description Verifies the signature of `message`, using the supplied pair
  */
-export default function schnorrkelVerify (message: Uint8Array, signature: Uint8Array, publicKey: Uint8Array): boolean {
-  return sr25519Verify(signature, message, publicKey);
+export default function schnorrkelVerify (message: Uint8Array | string, signature: Uint8Array | string, publicKey: Uint8Array | string): boolean {
+  const messageU8a = u8aToU8a(message);
+  const signatureU8a = u8aToU8a(signature);
+  const publicKeyU8a = u8aToU8a(publicKey);
+
+  return sr25519Verify(signatureU8a, messageU8a, publicKeyU8a);
 }